fix bug

+++ loss_func.py
@@ -94,7 +94,7 @@ def multiScale_loss(x,x_):
     #l3 = abs(1-vector_x.dot(vector_x_)/(torch.sqrt(vector_x.dot(vector_x))*torch.sqrt(vector_x_.dot(vector_x_))))
     l3 = (1-abs(torch.cosine_similarity(x.view(x.shape[0],-1),x_.view(x.shape[0],-1)))).mean()
 
-    l4 = loss_ce(x,x_)
+    l4 = loss_ce(x.long(),x_.long())
 
     l5 = loss_lpips(x,x_)
